New interface for v0.20
This is a big interface change where we no longer return the expected model output as the last column of the shap_values array. This makes KernelExplainer and TreeExplainer consistent with DeepExplainer, and also now requires us to pass `explainer.expected_value` to `force_plot` calls (since the base value is no longer the last column).

This should make everything much clearer in the long run, since now the shape of the returned shap_values matrix is the exact same as the passed input matrix.