Fixes error with XGBoost 1.1
In the new XGBoost (1.1), `.save_raw()` output is slightly different, which means that `TreeExplainer` doesn't work. See issue #1215. 
The build is currently failing for this reason, with all XGBoost tests failing.

The fix here is to just remove the prefix 'binf' from the model buffer, which as been added in the new version (either intentionally or accidentally - I can't find a PR in XGBoost which mentions this change.)

Fixes #1215