Critical Fix: backward permutation of PermutationExplainer unused
PermutationExplainer evaluates the model on both the forward and backward permutations but the output of the backward permutation is never used yielding an evaluation of the model of (up to) twice the amount of data needed and a potentially drastic slowdown of the model agnostic explainer. This fix includes the backward permutation in the SHAP value estimates.