Update _monitoring.py
If the 'inc' is greater than or equal to len(ys)-inc, the subsequent for loop does not run. Hence, the list pvals remains empty. As a result, np.min(pvals) fails resulting in failure of the entire monitoring_plot function. So, I corrected this error by allowing the user to choose 'inc' value at runtime by passing it as an argument to the function. Set the default value as 50 for backward compatibility. If (inc < len(ys)-inc), the t-test is skipped and a Warning is produced to prevent the failure of function of monitoring_plot.