ENH: faster non-tree KernelExplainer
* Fixes gh-3943.

* Cythonizing one of the inner loops used by a non-tree
`KernelExplainer` is about 30 % faster on the small `sklearn`
`SVC` test case in the matching issue. I'm not sure
if you'll accept a new dependency on Cython or ask
for a rewrite using the Python C-API (looks like you have
some limited use of the Python C-API at the moment).

* A standard `pip install -v .` followed by `pytest --import-mode=append`
(based on your apparent preferred test incantation) does pass the
test suite locally on this branch.

* I don't think I saw standard `asv` performance benchmarks
in the codebase, but let me know if I've missed them.