ENH: support rng for summary_plot (#3945)
* ENH: support rng for summary_plot

* Some folks on my team are having issues writing tests
downstream of `shap` because `summary_plot` mutates
the NumPy legacy random machinery global state, which is
hard to work around in a large testsuite. It would be
great if a heavily-used project in the ML space like `shap`
could start to adopt support for local random `Generator`
objects per the community document at:
https://scientific-python.org/specs/spec-0007/
This PR is a lot cruder than that approach, but appears
to allow preservation of the current global state behavior
while allowing the option of using the modern non-global
approach for downstream developers that need it.

* This patch adds a regression test that fails when not
using the optional `rng` argument, and passes when
it is used to scope to a local random state. If there
is a preference for using the full SPEC 7 approach
(and that may very well be the best idea), I may ask
one of my team members to help out a bit, since they
will benefit from it.

* The full testsuite passed locally via `pytest --import-mode=append`

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* MAINT: PR 3945 revisions

* Ignore some `mypy` false positives in
  `test_summary_plot_seed_insulated`

* Switch from `rng` to `seed` argument name for `summary_legacy()`
based on reviewer feedback, to leave an easier route open to SPEC 7.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* MAINT: PR 3945 revisions

* Based on reviewer feedback, embrace SPEC 7
and the new `rng` argument for `summary_legacy()`.
Raise a `FutureWarning` for usage of global random seed.

* Expand the testing of `summary_legacy()` to enforce
more behavioral aspects of SPEC 7.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* MAINT: PR 3945 revisions

* `summary_legacy()` docstring was updated to describe the
new `rng` parameter.

* ignore static typing check for `_bit_generator._seed_seq`

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>