ENH: faster non-tree KernelExplainer (#3944)
* ENH: faster non-tree KernelExplainer

* Fixes gh-3943.

* Cythonizing one of the inner loops used by a non-tree
`KernelExplainer` is about 30 % faster on the small `sklearn`
`SVC` test case in the matching issue. I'm not sure
if you'll accept a new dependency on Cython or ask
for a rewrite using the Python C-API (looks like you have
some limited use of the Python C-API at the moment).

* A standard `pip install -v .` followed by `pytest --import-mode=append`
(based on your apparent preferred test incantation) does pass the
test suite locally on this branch.

* I don't think I saw standard `asv` performance benchmarks
in the codebase, but let me know if I've missed them.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* add kernel lib.pyi

* add kernel lib.pyi

* add stubs

* add mypy.ini

* add mypy stubs config

* compile cython extensions before building docs

* install setuptools in readthedocs

* MAINT: PR 3944 revisions

* Add missing Cython dependency to `.readthedocs.yml`

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Tobias Pitters <31857876+CloseChoice@users.noreply.github.com>
Co-authored-by: Tobias Pitters <tobias.pitters@gmail.com>